Remove execute function in favor of HashMap
The HashMap stores associations between integer opcodes and method
references using lambdas. This is much better than the large, ugly
switch block that did this before.

The SICmulator class now handles the ArrayOutOfBoundsExceptions too
in order to catch errors with addressing, and remove a lot of excessive
code in the ALU class.